move Components directory
