+++ app/Components/Cards/Greetings.tsx
