+++ app/Components/Cards/MemoCard.tsx
