+++ app/Components/Forms/Memo/Forms/AudioMemoCreateForm.tsx
@@ -1,14 +1,12 @@
 import { useState } from "react";
 import axios from "axios";
 import { useRouter } from "next/navigation";
-import { useFormStatus } from "react-dom";
 
-import { ArrowRightCircle } from "@/Components/Icons";
-import { Button } from "@/Components/Button";
+import { Button } from "@/app/Components/Button";
 import { PostMemoButtonProps } from "@/Lib/Types";
 
-export default function ImageMemo() {
-  const [image, setImage] = useState<File | null>(null);
+export default function AudioMemo() {
+  const [audio, setAudio] = useState<File | null>(null);
   const [caption, setCaption] = useState<string>("");
   const [error, setError] = useState<string | null>(null);
   const [waiting, setWaiting] = useState(false);
@@ -21,11 +19,11 @@ export default function ImageMemo() {
   const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const uploadedFile = e.target.files?.[0];
     if (uploadedFile) {
-      setImage(uploadedFile);
+      setAudio(uploadedFile);
     }
   };
 
-  const handleCreateImageMemo = async () => {
+  const handleCreateAudioMemo = async () => {
     setWaiting(true);
     const accessToken = localStorage.getItem("accessToken");
     if (!accessToken) {
@@ -34,7 +32,7 @@ export default function ImageMemo() {
       setWaiting(false);
       return;
     }
-    if (!image) {
+    if (!audio) {
       setError("Image required for image memos.");
       setWaiting(false);
       return;
@@ -43,10 +41,10 @@ export default function ImageMemo() {
     try {
       const formData = new FormData();
       formData.append("caption", caption);
-      formData.append("memoFile", image);
+      formData.append("memoFile", audio);
 
       const response = await axios.post(
-        `${process.env.NEXT_PUBLIC_API_BASE_URL}/memo/image`,
+        `${process.env.NEXT_PUBLIC_API_BASE_URL}/memo/audio`,
         formData,
         {
           headers: {
@@ -71,7 +69,7 @@ export default function ImageMemo() {
           value={caption}
           onChange={handleInputChange}
           rows={2}
-          placeholder="What's this image about?"
+          placeholder="What's this audio about?"
         />
       </div>
       <div className="mb-6">
@@ -81,10 +79,10 @@ export default function ImageMemo() {
   rounded-md font-semibold text-xs uppercase tracking-widest hover:bg-gray-700 active:bg-gray-900 focus:outline-none 
   focus:border-gray-900 focus:ring ring-gray-300 disabled:opacity-25 transition ease-in-out duration-150"
         >
-          Select image
+          Select audio
           <input
             id="memoFile"
-            accept="image/*"
+            accept="audio/*"
             className="hidden"
             type="file"
             onChange={handleFileChange}
@@ -93,8 +91,8 @@ export default function ImageMemo() {
       </div>
       <div className="flex flex-col w-full">
         {error && <p className="text-red-500">{error}</p>}
-        <PostImageMemoButton
-          onPost={handleCreateImageMemo}
+        <PostAudioMemoButton
+          onPost={handleCreateAudioMemo}
           waitingStatus={waiting}
         />
       </div>
@@ -102,7 +100,7 @@ export default function ImageMemo() {
   );
 }
 
-function PostImageMemoButton({ onPost, waitingStatus }: PostMemoButtonProps) {
+function PostAudioMemoButton({ onPost, waitingStatus }: PostMemoButtonProps) {
   return (
     <Button
       className="flex justify-center mt-4 w-full text-white"
