+++ app/Components/Lists/FindFollowingList.tsx
@@ -1,4 +1,4 @@
-import SearchBar from "@/Components/Forms/SearchBar";
+import SearchBar from "@/app/Components/Forms/SearchBar";
 
 function FindFollowingList() {
   return <SearchBar />;
