+++ app/Components/Lists/FollowingMemosList.tsx
@@ -1,12 +1,13 @@
 "use client";
 
-import { Memo } from "@/Lib/Types";
-import MemoCard from "@/Components/Cards/MemoCard";
 import axios from "axios";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
-export default function AllMemosList() {
+import { Memo } from "@/Lib/Types";
+import MemoCard from "@/app/Components/Cards/MemoCard";
+
+export default function FollowingMemosList() {
   const [memos, setMemos] = useState<Memo[]>([]);
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
@@ -25,7 +26,7 @@ export default function AllMemosList() {
     }
     try {
       const response = await axios.get<Memo[]>(
-        `${process.env.NEXT_PUBLIC_API_BASE_URL}/memo/all?page=${pageNumber}&pageSize=5`,
+        `${process.env.NEXT_PUBLIC_API_BASE_URL}/memo/feed?page=${pageNumber}&pageSize=5`,
         {
           headers: { Authorization: `Bearer ${accessToken}` },
         }
