+++ app/explore/page.tsx
@@ -1,4 +1,4 @@
-import ExplorerTabs from "@/Components/Tabs/ExplorerTabs";
+import ExplorerTabs from "@/app/Components/Tabs/ExplorerTabs";
 
 export default async function ExplorePage() {
   return (
