+++ app/home/page.tsx
@@ -1,4 +1,4 @@
-import HomeTabs from "@/Components/Tabs/HomeTabs";
+import HomeTabs from "@/app/Components/Tabs/HomeTabs";
 
 export default async function DashboardPage() {
   return (
